% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercnv_heatmap_visualization.R
\name{plot_complex_heatmap}
\alias{plot_complex_heatmap}
\title{Plot Complex Heatmap for inferCNV Output}
\usage{
plot_complex_heatmap(
  infercnv_obj,
  output_dir,
  heatmap_thresholds_file,
  dendrogram_file,
  subcluster = FALSE,
  obs_class = "class",
  annotation_name = NULL,
  annotation_df = NULL,
  output_format = "png",
  output_name = "complex_heatmap",
  class_cols = NULL,
  anno_cols = NULL,
  ref_pie_chart = FALSE,
  sample_order = NULL
)
}
\arguments{
\item{infercnv_obj}{An inferCNV object created by the \code{infercnv::create_infercnv_object} and processed by \code{run()} or \code{run_infercnv()}.}

\item{output_dir}{Directory where the output heatmap PDF will be saved.}

\item{heatmap_thresholds_file}{Path to the inferCNV-generated \code{heatmap_thresholds.txt} file.}

\item{dendrogram_file}{A Newick file for dendrogram, can be either a single tree (\code{phylo}) or multiple trees (\code{multiPhylo}).}

\item{subcluster}{Change to true if subclustering was used in the inferCNV run.}

\item{obs_class}{Character string indicating how observations are annotated. Example \code{"Clone"}, \code{"Histology"}.}

\item{annotation_name}{Character string indicating what additional annotation is desired. Example \code{"Histology"}.}

\item{annotation_df}{A data frame or path to a TSV file with additional annotation information (annotation should be in the first column). Required if annotation_type is used.}

\item{output_format}{File format for the saved heatmap: \code{"png"} (default) or \code{"pdf"}.}

\item{output_name}{File name for the saved heatmap. Default is \code{"complex_heatmap"}. Extension is added automatically!}

\item{class_cols}{Optional. Named list or vector of colors for classes groups within object. Format: \code{c("A" = "#hex", ...)}.}

\item{anno_cols}{Optional. Named list or vector of colors for additional annotation groups. Format: \code{c("A" = "#hex", ...)}.}

\item{ref_pie_chart}{Optional. Adds a section distribution pie chart of the reference}

\item{sample_order}{Optional. If a specific sample or section order is desired in the legend.}
}
\value{
A image or pdf file is saved to \code{output_dir}, visualizing CNV patterns across references and observations with dendrograms and annotations.
}
\description{
This function generates a heatmap using the ComplexHeatmap package to visualize inferCNV results,
supporting reference and observation groups, one additional annotation, and user-defined color schemes.
}
\details{
Barcodes needs to be formatted in the following way: pt03_AP1_CAATGGATACGCTCGA.1 patient_section_10xBarcodeID.1 with Patient ID being 4 characters long.
\itemize{
\item For \code{dendrogram_file}, a single tree (\code{phylo}) will apply a global dendrogram, while a multiple tree (\code{multiPhylo}) input will order spots by original grouping.
\item The function will auto-detect and adjust for reference presence.
\item Colors are automatic assigned but can be overridden for class and anno.
\item Chromosome and CNV coloring is derived from inferCNV settings and thresholds.
\item Reference and query heatmap size will scale to relative sizes but has an lower and upper limit of 10\% and 25\%.
\item If subcluster is used during inferCNV run and set to TRUE, no additional annotation is needed. If additional annotations is provided using annotation_name and annotation_df these will be used instead of the ones stored in the inferCNV object
}
}
\examples{
plot_complex_heatmap(
  infercnv_obj = infercnv_final,
  output_dir = "results/",
  heatmap_thresholds_file = "results/infercnv.heatmap_thresholds.txt",
  dendrogram_file = "results/infercnv.observations_dendrogram.txt",
  obs_class = "Clone",
  annotation_name = "Histology",
  annotation_df = "metadata/histologies.tsv",
  class_cols = c("A" = "#FF0000", "B" = "#0000FF"),
  anno_cols = c("A" = "#E41A1C", "B" = "#377EB8"),
  ref_pie_chart = TRUE
)

# Subclusters from inferCNV
plot_complex_heatmap(
  infercnv_obj = infercnv_final,
  output_dir = "results/",
  heatmap_thresholds_file = "results/infercnv.heatmap_thresholds.txt",
  dendrogram_file = "results/infercnv.observations_dendrogram.txt",
  obs_class = "Histology",
  subcluster = TRUE,
  ref_pie_chart = TRUE
)
}
